Refactor ActivitiesRepository to own RealmUserChallengeActions queries
