+++ app/src/test/java/org/ole/planet/myplanet/repository/ProgressRepositoryImplTest.kt
@@ -46,8 +46,7 @@ class ProgressRepositoryImplTest {
             databaseService,
             UnconfinedTestDispatcher(),
             dispatcherProvider,
-            { mockCoursesRepository },
-            { mockk(relaxed = true) }
+            { mockCoursesRepository }
         ), recordPrivateCalls = true)
     }
 
@@ -309,25 +308,20 @@ class ProgressRepositoryImplTest {
 
     @Test
     fun testHasUserCompletedSync() = testScope.runTest {
-        val activitiesRepo = mockk<ActivitiesRepository>()
-        val localRepository = ProgressRepositoryImpl(
-            databaseService,
-            UnconfinedTestDispatcher(),
-            dispatcherProvider,
-            { mockCoursesRepository },
-            { activitiesRepo }
-        )
+        coEvery {
+            repository invoke "count" withArguments listOf(org.ole.planet.myplanet.model.RealmUserChallengeActions::class.java, any<Function1<io.realm.RealmQuery<org.ole.planet.myplanet.model.RealmUserChallengeActions>, Unit>>())
+        } returns 1L
 
-        coEvery { activitiesRepo.hasUserCompletedSync("user1") } returns true
-
-        val result = localRepository.hasUserCompletedSync("user1")
+        val result = repository.hasUserCompletedSync("user1")
         advanceUntilIdle()
 
         assertEquals(true, result)
 
-        coEvery { activitiesRepo.hasUserCompletedSync("user1") } returns false
+        coEvery {
+            repository invoke "count" withArguments listOf(org.ole.planet.myplanet.model.RealmUserChallengeActions::class.java, any<Function1<io.realm.RealmQuery<org.ole.planet.myplanet.model.RealmUserChallengeActions>, Unit>>())
+        } returns 0L
 
-        val result2 = localRepository.hasUserCompletedSync("user1")
+        val result2 = repository.hasUserCompletedSync("user1")
         advanceUntilIdle()
 
         assertEquals(false, result2)
