+++ app/src/test/java/org/ole/planet/myplanet/repository/UserRepositoryImplTest.kt
@@ -88,8 +88,7 @@ class UserRepositoryImplTest {
             context,
             configurationsRepository,
             appScope,
-            dispatcherProvider,
-            mockk(relaxed = true)
+            dispatcherProvider
         )
     }
